+++ docs/changes.rst
@@ -5,7 +5,6 @@ Changelog
 ======
 * The memcached backend has been removed
 * Keys have to be provided as unicode strings
-<<<<<<< HEAD
 * Values have to be provided as bytes (python 2) or as str (python 3)
 * keys() and iter_keys() provide a parameter to iterate just over all keys with a given prefix
 * Added :class:`simplekv.CopyMixin` to allow access to copy operations to
@@ -16,7 +15,8 @@ Changelog
   :class:`~simplekv.decorator.URLEncodeKeysDecorator`
 * Added a Microsoft Azure Blob Storage backend:
   :class:`~simplekv.net.azurestore.AzureBlockBlobStore`
-* Added :class:`~simplekv.ExtendedKeyspaceMixin` which allows slashes and spaces in key names
+* Added :class:`~simplekv.contrib.ExtendedKeyspaceMixin` which allows slashes and spaces in key names
+  This mixin is experimental, unsupported and might not work with all backends.
 
 
 0.10.0